not removing all query parameters once a subparameter has been emptied
